first pass at splitting up eloquent
